Rename test-utils to test Remove done TODOs Move some tests to handleUpdates
