+++ src/Popup/sendImages.test.ts
@@ -1,4 +1,4 @@
-import { asMockedFunction, mockChrome } from '../test-utils';
+import { asMockedFunction, mockChrome } from '../test';
 
 declare var global: any;
