+++ src/test.ts
@@ -1,4 +1,3 @@
-// TODO: Rename to avoid potential conflict with `utils.js`
 import { mockRecursivePartial } from 'sneer';
 
 export const mockChrome = () =>
